requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
streamlit==1.31.1
fastapi==0.109.2
uvicorn==0.27.1
torch==2.2.0
transformers==4.37.2
tensorflow==2.15.0
numpy==1.26.4
yfinance==0.2.36
yahooquery==2.3.7
finnhub-python==2.4.19
pandas==2.2.0
plotly==5.18.0
scikit-learn==1.4.0
requests==2.31.0
python-multipart==0.0.9
python-dotenv==1.0.0
pydantic==2.6.1
